requirements.txt
asgiref==3.3.1
cached-property==1.5.2
cffi==1.14.4
cryptography==3.3.1
dj-database-url==0.5.0
Django==3.1.5
django-cors-headers==3.6.0
django-heroku==0.3.1
django-rest-knox==4.1.0
django-rest-passwordreset==1.1.0
django-summernote==0.8.11.6
django-url-filter==0.3.15
djangorestframework==3.12.2
enum-compat==0.0.3
gunicorn==20.0.4
Pillow==8.1.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
python-decouple==3.4
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
whitenoise==5.2.0
